reimplementing the 8-router IS-IS testbed w/o dynamic routing daemons
The 8-router IS-IS testbed was carefully reconstructed while ensuring
that all dependencies on Zebra and IS-IS dynamic routing daemons were
eliminated. The objective of this reconstruction was to establish a more
controlled and reproducible self-contained testing environment.

In order to enhance configurability and streamline the setup process,
the provided script incorporates adjustable setting variables. These
variables facilitate the implementation of specific SRv6 policies, which
are essential for establishing Layer 3 Virtual Private Networks (L3VPNs)
between host nodes.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>